Update GetDropboxPath
It seems that Dropbox does not use the files config.db and host.db as it
has before. Because of that GetDropboxPath was unable to find the
folder.

Now it checks the DROPBOX_FILTER_PATH environment variable for the
correct path. If this variable is not available it falls back to the
default path in the user's home directory (HOMEDRIVE\HOMEPATH\Dropbox).